Moved week 8 code to a new package to avoid confusion and better support week 9 testing
